Updates scanner to better handle newlines
The handcrafted scanner attempted to codify some of the the complex
lookahead rules for newlines in scala. However, this only captured
"automatic" semicolons (multiple newlines) if one was found, and checked
that single newlines were not followed by a token which could start a
new statement. These changes do all of that, and also captures a single
newline as a token which can be used in the grammar.